+++ lib/src/bones_api_entity_db.dart
@@ -1101,22 +1101,26 @@ class DBEntityRepository<O extends Object> extends EntityRepository<O>
   FutureOr<dynamic> _ensureStoredImpl(
       o, Transaction? transaction, TransactionOperation? parentOperation) {
     if (transaction != null) {
-      var storedOp = transaction
-          .firstOperationWithEntity<TransactionOperationSaveEntity>(o);
+      var storeOp =
+          transaction.firstOperationWithEntity<TransactionOperationStore>(o);
 
-      if (storedOp != null) {
-        return storedOp.waitFinish().then((_) {
-          var id = getEntityID(storedOp.entity) ?? getEntityID(o);
+      if (storeOp != null) {
+        return storeOp.waitFinish(parentOperation: parentOperation).then((ok) {
+          var id = getEntityID(storeOp.entity) ?? getEntityID(o);
+          if (id == null && !ok) {
+            throw RecursiveRelationshipLoopError.fromTransaction(
+                transaction, storeOp, parentOperation, o);
+          }
           return id;
         });
       }
     }
 
-    return store(o, transaction: transaction);
+    return _storeImpl(o, transaction, parentOperation);
   }
 
   @override
-  FutureOr<bool> ensureReferencesStored(o,
+  FutureOr<bool> ensureReferencesStored(O o,
       {Transaction? transaction, TransactionOperation? operation}) {
     throw UnsupportedError("Relationships not supported for: $this");
   }
@@ -1136,8 +1140,8 @@ class DBEntityRepository<O extends Object> extends EntityRepository<O>
       Transaction? transaction}) {
     checkNotClosed();
 
-    var op = TransactionOperationCount(
-        name, operationExecutor, matcher, transaction);
+    var op = TransactionOperationCount(name, operationExecutor,
+        matcher: matcher, transaction: transaction);
 
     try {
       return repositoryAdapter.doCount(op,
@@ -1200,7 +1204,8 @@ class DBEntityRepository<O extends Object> extends EntityRepository<O>
     var canPropagate = hasReferencedEntities(resolutionRulesResolved);
 
     var op = TransactionOperationSelect(
-        name, canPropagate, operationExecutor, matcher, transaction);
+        name, canPropagate, operationExecutor, matcher,
+        transaction: transaction);
 
     try {
       return repositoryAdapter.doSelectByID<O?>(op, id, preFinish: (results) {
@@ -1232,7 +1237,8 @@ class DBEntityRepository<O extends Object> extends EntityRepository<O>
     var canPropagate = hasReferencedEntities(resolutionRulesResolved);
 
     var op = TransactionOperationSelect(
-        name, canPropagate, operationExecutor, matcher, transaction);
+        name, canPropagate, operationExecutor, matcher,
+        transaction: transaction);
 
     try {
       return repositoryAdapter.doSelectByIDs<O>(op, ids, preFinish: (results) {
@@ -1257,7 +1263,8 @@ class DBEntityRepository<O extends Object> extends EntityRepository<O>
     var canPropagate = hasReferencedEntities(resolutionRulesResolved);
 
     var op = TransactionOperationSelect(
-        name, canPropagate, operationExecutor, matcher, transaction);
+        name, canPropagate, operationExecutor, matcher,
+        transaction: transaction);
 
     try {
       return repositoryAdapter.doSelectAll<O>(op, preFinish: (results) {
@@ -1703,20 +1710,24 @@ class DBEntityRepository<O extends Object> extends EntityRepository<O>
   }
 
   @override
-  FutureOr<dynamic> store(O o, {Transaction? transaction}) {
+  FutureOr<dynamic> store(O o, {Transaction? transaction}) =>
+      _storeImpl(o, transaction, null);
+
+  FutureOr<dynamic> _storeImpl(
+      O o, Transaction? transaction, TransactionOperation? parentOperation) {
     checkNotClosed();
 
     checkEntityFields(o);
 
     if (isStored(o, transaction: transaction)) {
-      return _update(o, transaction, true);
+      return _update(o, transaction, parentOperation, true);
     }
 
     var canPropagate = hasReferencedEntities(
         resolveEntityResolutionRules(EntityResolutionRules.instanceAllEager));
 
-    var op = TransactionOperationStore(
-        name, canPropagate, operationExecutor, o, transaction);
+    var op = TransactionOperationStore(name, canPropagate, operationExecutor, o,
+        transaction: transaction, parentOperation: parentOperation);
 
     try {
       var idFieldsName = entityHandler.idFieldName(o);
@@ -1739,13 +1750,14 @@ class DBEntityRepository<O extends Object> extends EntityRepository<O>
     }
   }
 
-  FutureOr<dynamic> _update(
-      O o, Transaction? transaction, bool allowAutoInsert) {
+  FutureOr<dynamic> _update(O o, Transaction? transaction,
+      TransactionOperation? parentOperation, bool allowAutoInsert) {
     var canPropagate = hasReferencedEntities(
         resolveEntityResolutionRules(EntityResolutionRules.instanceAllEager));
 
     var op = TransactionOperationUpdate(
-        name, canPropagate, operationExecutor, o, transaction);
+        name, canPropagate, operationExecutor, o,
+        transaction: transaction, parentOperation: parentOperation);
 
     var idFieldsName = entityHandler.idFieldName(o);
     var id = entityHandler.getID(o);
@@ -1841,7 +1853,8 @@ class DBEntityRepository<O extends Object> extends EntityRepository<O>
         hasReferencedEntities(resolveEntityResolutionRules(null));
 
     var op = TransactionOperationDelete(
-        name, canPropagate, operationExecutor, matcher, transaction);
+        name, canPropagate, operationExecutor, matcher,
+        transaction: transaction);
 
     try {
       return repositoryAdapter.doDelete(op, matcher,